feat: mobile search header 추가
